feat: 출석체크 완료 시 데이터 변경 반영하기 (temp)#32
